[JA-5] - sayı if-else yapısıyla kontrol edildi

+++ src/ArithmeticOperations.java
@@ -4,10 +4,10 @@ public class ArithmeticOperations {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Sayı giriniz: ");
-        float number1 = sc.nextInt();
+        float number1 = sc.nextFloat();
 
         System.out.println("Sayı giriniz: ");
-        float number2 = sc.nextInt();
+        float number2 = sc.nextFloat();
 
         float sum = number1 + number2;
         System.out.println("Sayıların toplamı: " + sum);
